Move Sketches test utils to sketches module
