+++ java/compaction/compaction-job-execution/pom.xml
@@ -125,6 +125,13 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
+            <artifactId>sketches</artifactId>
+            <version>${project.parent.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
             <artifactId>compaction-job-creation</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
