+++ java/sketches/src/test/java/sleeper/sketches/testutils/AssertQuantiles.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.testutils;
+package sleeper.sketches.testutils;
 
 import org.apache.datasketches.quantiles.ItemsSketch;
